image_rec/requirements.txt
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.5.0
google-auth==1.5.1
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.5
requests==2.20.0
rsa==4.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
